Fix cZone::distance signature

+++ src/cZone.cpp
@@ -214,8 +214,8 @@ namespace pup
     }
 
     float cZone::distance(
-        int x1, int y1,
-        int x2, int y2,
+        float x1, float y1,
+        float x2, float y2,
         eDistMethod m)
     {
         switch (m)
