+++ src/cZone.h
@@ -75,8 +75,8 @@ namespace pup
                 m);
         }
         static float distance(
-            int x1, int y1,
-            int x2, int y2,
+            float x1, float y1,
+            float x2, float y2,
             eDistMethod m = eDistMethod::manhatten);
 
     private:
